requirements.txt
absl-py==2.3.1
astunparse==1.6.3
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
dm-tree==0.1.9
docstring_parser==0.17.0
einops==0.8.1
etils==1.13.0
flatbuffers==25.2.10
fsspec==2025.9.0
gast==0.6.0
google-pasta==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.75.0
h5py==3.14.0
idna==3.10
immutabledict==4.2.1
importlib_resources==6.5.2
joblib==1.5.2
keras==3.11.3
libclang==18.1.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.1.0
numpy==1.26.4
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.2
pillow==11.3.0
promise==2.3
protobuf==4.25.8
psutil==7.1.0
pyarrow==21.0.0
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
rich==14.1.0
scikit-learn==1.7.2
scipy==1.16.2
shellingham==1.5.4
simple-parsing==0.1.7
six==1.17.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.1
tensorflow-datasets==4.9.9
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.17.2
tensorflow-metal==1.2.0
termcolor==3.1.0
threadpoolctl==3.6.0
toml==0.10.2
tqdm==4.67.1
typer==0.17.4
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.17.3
zipp==3.23.0
